backend/requirements.txt
# Backend dependencies for Urdu Character Recognition API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.2
opencv-python==4.8.1.78
pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.2
